Replace faulty OVIS_EXEC_SCRIPTS macro with direct AC_CONFIG_FILES
The OVIS_EXEC_SCRIPTS macro probably never really worked correctly
to set executable permissions on the scripts. This is likely why
AC_CONFIG_COMMANDS was used to seemingly redundantly set executable
flags on the scripts.  It wasn't really redundant because the chmod
in OVS_EXEC_SCRIPTS wasn't working.

In fact, none of the AC_CONFIG_FILES calls under OVIS_EXEC_SCRIPTS
were fulling working, which is why autoconf didn't know to include
the *.in files in the dist, and OVIS_EXEC_SCRIPTS tried to patch
over that by setting up explicit EXTRA_DISTS. The cause of the
disfunction is probably the use of shell variables in the first
option to AC_CONFIG_FILES (autoconf calls this the "tags" field).
Tags are explicitly encouraged to be _literals_ in the autoconf
manual, and using shell variables is called out as something to
avoid.

By eliminating OVIS_EXEC_SCRIPTS and calling AC_CONFIG_FILES
directly for each script, we can switch to using literals tags,
and all of the issues fall away.

Fixing this revealed some other errors in dist usage that are
also cleaned up.